+++ core/server/controllers/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    preview: require('./preview'),
+    single: require('./single')
+};
